feat: Hook everything up
This commit actually contains many things. The most obvious change is
that all the other code is actually used to generate a response! Other
than that there was also a need to set up CORS. This commit also
implements the route in such a way that if YT categorization is
available the frames are never downloaded and classified leading to a
massive speed boost

Closes: #28